Component refactor
